requirements.txt
pandas~=2.2.3
numpy~=1.26.4
scikit-learn~=1.6.1
xgboost~=3.0.0
lightgbm~=4.6.0
catboost~=1.2.7
matplotlib~=3.10.1
flask~=3.1.0
mysql-connector-python~=9.3.0
bcrypt
flask_cors~=5.0.1
joblib~=1.4.2
spglib~=2.6.0
streamlit~=1.44.1
requests~=2.32.3
seaborn~=0.13.2
pillow~=11.1.0
faiss-cpu~=1.11.0
sentence-transformers~=4.1.0
datasets~=3.6.0
transformers~=4.52.4
torch~=2.7.1